removed_types,support multi-modal inputs
