+++ clarifai/client/app.py
@@ -6,10 +6,7 @@ from google.protobuf.json_format import MessageToDict
 
 from clarifai.client.base import BaseClient
 from clarifai.client.dataset import Dataset
-from clarifai.client.input.audio import Audio
-from clarifai.client.input.image import Image
-from clarifai.client.input.text import Text
-from clarifai.client.input.video import Video
+from clarifai.client.input import Inputs
 from clarifai.client.lister import Lister
 from clarifai.client.model import Model
 from clarifai.client.workflow import Workflow
@@ -169,6 +166,13 @@ class App(Lister, BaseClient):
 
     return Workflow(**kwargs)
 
+  def inputs(self,):
+    """Returns an Input object.
+    Returns:
+        Inputs: An input object.
+    """
+    return Inputs(self.user_id, self.id)
+
   def delete_dataset(self, dataset_id: str) -> None:
     """Deletes an dataset for the user.
     Args:
@@ -202,34 +206,6 @@ class App(Lister, BaseClient):
     if response.status.code != status_code_pb2.SUCCESS:
       raise Exception(response.status)
     self.logger.info("\nWorkflow Deleted\n%s", response.status)
-
-  def image(self) -> Image:
-    """Returns an Image object.
-    Returns:
-        Image: An Image object.
-    """
-    return Image(self.user_id, self.id)
-
-  def video(self) -> Video:
-    """Returns a Video object.
-    Returns:
-        Video: A Video object.
-    """
-    return Video(self.user_id, self.id)
-
-  def audio(self) -> Audio:
-    """Returns an Audio object.
-    Returns:
-        Audio: An Audio object.
-    """
-    return Audio(self.user_id, self.id)
-
-  def text(self) -> Text:
-    """Returns a Text object.
-    Returns:
-        Text: A Text object.
-    """
-    return Text(self.user_id, self.id)
 
   def __getattr__(self, name):
     return getattr(self.app_info, name)
